[BUG] AudioPolicyManager: prevent unecessary force routing on SwOutput
At the begining/end of call, routing for all SwOutput is evaluated.
In case of using bridge, either the output involved in the SwBridge or
the output involved in the HwBridge for volume will be forced to be rerouted
on same device. As a result, AudioHAL patch already created will not be
released but recreated. This leak prevents AudioHAL implementers to properly
take into account their unique handle for routing management.

This CL filters this unecessary force routing

Bug: 187173302

Test: CSV & VoIP calls

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ia15f137aa543ef022cac20c51045f44e423b9319