[BUG] AudioPolicyManager: fix bridges routing & volume issues
-When using either SwBridge or HwBridge, a SwOutput is always involved.
SwBridge: for playback, HwBridge: for volume control (it hosts the
Source Client).

-Prevent unecessary routing on SwOutput involved in call (current
check assumes primaryOutput is used)

-Volume Issue on Combo Hw/SwBrige on same output:
If an Output descriptor is used for volume control while an
Audio Source is active and realized through an Hw Bridge (Device
to Device audio patch), this output is declared as active but
is not really routed. Laterly, it will prevent any client to be
effectively routed on this output.

-Volume using HwGain:
2 clients following different volume group are connected to
a SwOutput. This output is routed to a device using Hw Gain.
The volume applied on the port if the one of the higher priority
strategy (order matters in the engine configuration file).
When the client with highest priority ends, the volume is not
reconsidered and is kept at the value of stopped output.
This CL forces to reapply the volume if SwOutput has still an
active client with different volume group and if routed device
is controling volume through Hw Gain.

Bug: 187173302

Test: build & audio UC involving bridges
(functional on emulator):
adb shell ./data/AudioPolicyEmulatorTests --gtest_filter=*AudioPatchTest*
adb shell ./data/AudioPolicyEmulatorTests --gtest_filter=*AudioSourceTest*

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I1198a4f9227f69e60db9f1a7925cbb3c93e3654c