C2 Vts: Avoid using strings to pass integer and bool values
In order to display parameters as index, tests were currently
using android::hardware::PrintInstanceTupleNameToString which
required all the parameters to be of the same type.
This limitation required passing integers and booleans as strings
and had to be converted back and forth.

This is now fixed by implementing a custom
PrintInstanceTupleNameToString that handles mixed variable types
in a tuple.

Bug: 155354751
Test: atest VtsHalMediaC2V1_0TargetAudioDecTest \
      VtsHalMediaC2V1_0TargetAudioEncTest \
      VtsHalMediaC2V1_0TargetVideoDecTest \
      VtsHalMediaC2V1_0TargetVideoEncTest \
      VtsHalMediaC2V1_0TargetComponentTest \
      VtsHalMediaC2V1_0TargetMasterTest

Change-Id: I49e63cd03ad7f5f691f6e8f32f40054f8c1e037e