CCodec: Control the inputs to avoid pipeline overflow
When the pipeline is almost full, do not indicate to app
about free input buffers if the pipeline is expected to
overflow in case the app queues more input buffers.

This is avoid watchdog timeout if app pauses the pipeline
immediately after pipeline overflows

Use this only for HW decoder.

Change-Id: If9eb24b0dcbe0a6e52da1d457c8af4b7d75aea4b