Fix C2StreamUsageTuning work for Linear Block model.
Because C2FieldSupportedValues has a default flag value 0, it causes
CCodec2 always to get minUsage with 0. So, Codec2 has no opportunity
to update the memory usage for fetchLinearBlock. Here, we change the
flow to let CCodec2 choose the second supported value if the supported
value number is larger than 1. So, Codec2 can decide the memory usage
for fetchLinearBlock to get the required memory attributes.

[Test Report]
Vendor DTV test: OK

Bug: 187972014
Change-Id: I691295330f061958bac3ebe80617c7c79a983420