reduce cpu MIPS when record
MPEG4Writer will full parse buffer to split NALs
no matter how many NALs in a buffer. This will cause
waste of resources when there is only single NAL in a buffer.

set a flag in buffer meta from component to writer to indicate
if there is only singal NAL in buffer, if yes, skip full parse
the buffer to reduce cpu MIPS.

Bug: 193761501
Test: manual, record a video with codec set single NAL flag, record success and MIPS reduced.

Change-Id: Ib9f9ae1afa8c55f95e3e81018169a7493052b16b