[IMPROVEMENT] AudioPolicyManager: optional release patch for Idle SwOutputs
AudioHAL implementers may claim about difficulty to track handle created
when creating the patch, as it may be never released. They need to implement
a garbage collector.
This CL adds uses a property to force or not the release of the Audio Patch
when the SwOuput is not used anymore.
Default is the legacy behavior.

Note that cold latency may be fewly altered as AudioHAL implementers
may unroute when a stream is stopped/paused to save power rather than on
release of the patch.

Test: build & basic audio

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Id2ada6f2d381690b75405cebe7ad4ac33f2b0d35