[BUG] AudioPolicyManager: fix volume on output using Hw Gain
2 clients following different volume group are connected to
a SwOutput. This output is routed to a device using Hw Gain.
The volume applied on the port if the one of the higher priority
strategy (order matters in the engine configuration file).

When the client with highest priority ends, the volume is not
reconsidered and is kept at the value of stopped output.

This CL forces to reapply the volume if SwOutput has still an
active client with different volume group and if routed device
is controling volume through Hw Gain.

Bug: 187173302

Test: build & boot & basic audio.

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Id6e601073b0c91f6de674d90174e073745b02fe2