[BUG] AudioPolicyManager: prevent unecessary force routing on SwOutput
At the begining/end of call, routing for all SwOutput is evaluated.
In case of using bridge, either the output involved in the SwBridge or
the output involved in the HwBridge for volume will be forced to be rerouted
on same device.

For AudioHAL < 6.0: this CL prevents patch leak will happen as
updateAudioPatch not present.
For AudioHAL >= 6.0: this CL is a matter of optimization.
As a result, AudioHAL patch already created will not be

Bug: 187173302

Test: CSV & VoIP calls

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ia15f137aa543ef022cac20c51045f44e423b9319