Dynamically enable ACodec logv by setprop
ACodec is a common part for media.
It's convenient to be able to enable logv dynamically via setprop,
especially in vendor platforms
that are sometimes hard to replace ".so", SQA test
or when we did not had source code to build.

Test:
YouTube,NetFlix and so on dynamic get logv OK

Bug: 138550887
Change-Id: Iead61c6b646773b6eaf379e7f1e191279a0e802b