[BUG] AudioPolicy: combo UC: mute may not work
When a SwOutput hosts a combo use case, if this output is routed
to a device using an HW Gain, the highest priority product strategy/
volume source will only be allowed for volume operation.
If muting the other volume source (adjustStreamVolume on AudioService
translated into setVolumeIndexForAttributes), the AudioPolicyManager
will block the volume change.

This CL fixes this issue by considering the mute and implement it
through Sw gain on AudioFlinger in case of concurrent UC on Output.

Test: audio smoke tests
Bug: 187173302

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I4742127d18befcfb6020717506ee68336a5714a7