Solve the problem of abnormal volume acquisition in some cases(2/4)
1. The boot animation sound uses the AUDIO_STREAM_ENFORCED_AUDIBLE type
to play, and the volume uses MUSIC instead of ENFORCED_AUDIBLE. If our
MUSIC level is 15 levels by default, the problem is not easy to find
because there is still sound, but if the maximum volume level is set to
150, the sound will be extremely low (the volume becomes 7/150 instead
of 7 /7)
2. "getAudioAttributesForStrategyWithLegacyStreamType" gets the default
value of the Attributes flag based on the new Attributes and the new
rules. This will cause the maximum volume acquisition to fail.

bug:195126766

Signed-off-by: lugandong <lugandong@xiaomi.com>
Change-Id: I17525a5edb96e95cbf88bec22ebd0ef1874de214