[TimedTextUnitTest] use shared libstagefright_foundation instead of static
libmedia.so, a dependency of TimedTextUnitTest, also depends on
libstagefright_foundation.so.  If we include libstagefright_foundation.a
when linking TimedTextUnitTest, upstream's lld linker will unnecessarily
fetch symbols from libstagefright_foundation.a to redefine undefined
symbols from libmedia.so.  Switching to libstagefright_foundation.so
instead does not increase the closure of shared libraries needed for
this test.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I4077a385f8c6233108f09a69260d0bda7a022099