Simplify MTP transcoding file size logic
There where 2 variables to represents the file size (`finalSize` & `fileLength`).
I think the intent was for both of them to have the same value, but sometime one is not updated.
`finalSize` is now removed.

If the transcoding branch is taken, but transcoding query fails, then the file size is the transcoded file (0?) but the fd is the original one.

Additionally a `stat` syscall was removed as the original file size is already
known from the `datablase->getObjectFilePath` call.

Bug: 190422448
Change-Id: Idb969681cd682fb01ff117a2913cdcceefb67ade