[BUG] AudioPolicyManager: fix port id alignment with JAVA layer
As device equality is based on {type,address} couple.
If a device is mounted, attached to a  module, an id is assigned.
if the device is unmounted, thus detached from module, id is resetted.
If this device is mounted again, thus attached to same module,
the routing may not have change (aka same output device for the concerned
SwOutput), setOutputDevice will not recreate/refresh the patch associated.
It may lead to misalignement with JAVA listAudioPort/listAudioPatch API
based on the mId to associated sources/sink of a patch to the port.

This CL ensures the patch is either removed (output is closed)
or updated (change of output device) depending if the
output used to reach the concerned device is still needed or not
after disconnection.

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ibfce32e2fea6dfb38afa0787ee1bddfc5fede0aa