stagefright: explicitly initialize variables
These are conditionally initialized in a way that one is ~always
initialized and the other isn't (depending on input). Though we have
compiler mitigations in place for this kind of thing, make the
initialization explicit since we'd prefer not to rely heavily on those.

Caught by the static analyzer:
> frameworks/av/media/libstagefright/ACodec.cpp:4823:12: warning: 2nd
function call argument is an uninitialized value
[clang-analyzer-core.CallAndMessage]
> frameworks/av/media/libstagefright/ACodec.cpp:4823:12: warning: 3rd
function call argument is an uninitialized value
[clang-analyzer-core.CallAndMessage]

Bug: None
Test: TreeHugger
Change-Id: I718275c89e0f5f291f0f755a0903fb68366b9954