C2 Decoders: Use P010 if supported on device for VP9 and AV1
Use P010 format for VP9 and AV1 10 bit decoding on devices where
that format is supported

P010, RGBA1010102 (for specific color aspects) and YV12 is the order
in which formats are tried.

AHardwareBuffer_isSupported() doesn't handle P010 format and hence
fetchGraphicBlock() call is used to determine which are formats are
supported.

Bug: 178229371
Test: atest CtsMediaV2TestCases -- --module-arg \
 CtsMediaV2TestCases:instrumentation-arg:codec-prefix:=c2.android.

Change-Id: I6503ecf49934f7d0fb8d8ffcf95e3aa4905741d8