[BUG] AudioPolicyManager: fix closure of output supporting multiple devices
CL:
https://android-review.googlesource.com/c/platform/frameworks/av/+/1863726
introduced a regression.

This CL improves the closure behavior of output supporting more than
on devices upon disconnection of one of them. This is particularily
impacting primary output since the tracking pointer is cleared, and will
not be re-assigned until one of the devices reachable by primary is reconnected,
regardless of still available devices reachable through this output.

This CL fixes this behavior by reconsidering the output supporting a
device which is disconnected. Output will either be rerouted if active
on the still available device(s). If not active, output will be routed
on first available device reachable from it. If none device available,
output can be safely closed.

It allows to prevent from misalignement with JAVA listAudioPort/listAudioPatch API
based on the mId to associated sources/sink of a patch to the port
by forcing the patch recreation/deletion and not maintaining
a patch involving a disconnected device.

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ibfce32e2fea6dfb38afa0787ee1bddfc5fede0aa