Fix BufferQueue issue
In the destructor, detachBuffer and cancelBuffer will be called. If two
decoders connect to the same surface, it may cause issue in following
situation:
a. The first decoder is doing detachBuffer or cancelBuffer
b. At the same time, the second decoder is doing dequeueBuffer from the
   same BufferQueue
c. If BufferQueue returns slot X to the second decoder by dequeueBuffer,
   and immediately after that, the first decoder calls detachBuffer
   using slot X, then the second decoder may get the same slot from
   dequeueBuffer.
d. The second decoder will encapsulate slot X into two different
   C2GraphicBlock, but they are the same buffer actually.
e. The content in the first C2GraphicBlock may be rewrite by the second
   C2GraphicBlock before it is sent to display.

Now that the first decoder has disconnected from the surface, there is
no need for it to call detachBuffer or cancelBuffer again. Skip this
function call to avoid affecting other operations to the same
BufferQueue.

Bug: 227121136
Test: test with TikTok
Change-Id: Ic6eb0995c02408d95cf64d18f4626825e5bee7e3