Introduce video peek legacy mode in tunnel mode.
Video Peek in Tunnel Mode is a MediaCodec feature introduced in
Android S that lets an app specify whether to display the first queued
frame before starting playback.

Before the feature was introduced, this behavior was unspecified, and
different decoders would behave differently. One unforeseen
consequence of the specification is that some of the previous
behaviors, although desirable, have been forbidden.

The purpose of this change is to introduce an escape hatch, dubbed
"legacy mode", which allows a decoder to revert to the previous state
of things by ignoring the video peek machinery. This legacy mode is
configured by default, and is opted out of as soon an app specifies
its preferred video peek configuration.

We intend to deprecate legacy mode in the feature, when we feel
confident all the existing desirable use cases are covered by well
specified decoder APIs.

Bug: 216283761
Test: $ atest DecoderTest
Change-Id: I22af113186f382137fc4e6ad9ab8b6f2ef4dca50