Support PCM encoding types for gapless playback
SkipCutBuffer removes audio encoder delay and encoder padding, but
assumes that the PCM encoding type is int 16 bit. As a result, silent
data remains for gapless playback or sound corruption occurs when the
padding frames are odd.
Fix the issue by supporting all PCM encoding types, int 8 bit, int 24
bit packed and int 32 bit as well as int 16 bit. The encoding format is
taken from the buffer output format, and if not specified, the default
int 16 bit is used.

Bug: 238281121
Test: playback a MP4 file with MPEG-H decoder outputting float.
Test: ran libstagefright_skipcutbuffer_fuzzer for 2hr.
Change-Id: Ic73359b6d35264c026cd503d92a81cd993f5c726