Do not saturate signal after gain amplification
Here within floating point arithmetic, saturation should not be done
even if the signal exceeds between -1.0 and 1.0.  In this
implemantation before this change, signal was once saturated and then
the gain was lowered, resulting in distorted sound in quite low level.
In this change, a new amplification function is introduced not to
saturate the signal here.
If necessary, saturation should be done in the conversion process from
floating point to PCM format.

Bug: XXX
Test: No distortion when playing music with EnvironmentalReverb
Change-Id: Ic96e697e3e91c3cbf06b35acebf886a1bb059b67