Fix FastMixer audio balance
This moves the balance channel configuration for the fast mixer
thread to the correct place AFTER the channel mask has been
updated. Previously the channel configuration was set to an invalid
value (some default - seems to be 2 stereo channels).

Test: Run `adb shell settings put system master_balance -1` while
running an app with the fast mixer enabled. Bug repro'd and fix
verified on a Meta Quest 2.

Change-Id: I3302386f91687afb7079f4d63dafc143742a412d