Fix FastMixer audio balance
This moves the balance channel configuration for the fast mixer thread to
the correct place AFTER the channel mask has been updated.  Previously the
channel configuration was set to an invalid value (some default - seems
to be 2 stereo channels).

Test: Run `adb shell settings put system master_balance -1` while
playing Beat Saber or Medal of Honor on a Meta Quest or Quest 2. Before
the patch it doesn't work, after the patch it correctly sets the audio
to left channel only.

Change-Id: I3302386f91687afb7079f4d63dafc143742a412d