Revert "MediaMetrics: Fix inaccurate attribution of thread device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bfa694e7-ec65-4a78-82dd-22ad6ebae0be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie8f2064afc6655706a340e89492db331e4972fa9