Revert "Spatializer: Send metrics when not created"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bfa694e7-ec65-4a78-82dd-22ad6ebae0be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09479d6f715bdadb32176d16cee83f50a225362a