Unit test to validate audio effect
The current test applies audio effect on a multi-tone signal. The
tones are selected according to the filter under test. For equalizer
effect, the center frequencies of all equalizer bands are chosen.
For bassboost effect, a bass frequency and speech tone is selected.
The output signal (effect applied signal) is captured and analysed.
This signal is expected to show the effects of filter characteristics

Bug:
Test: atest audioeffect_analysis

Change-Id: I848cd79cf1515be721e6e1b4d0240db0c762a91e