Re work audio test utils of audioclient test
audio_test_utils.cpp is using quite a lot of fixed constants. These
constants are used as timeouts. This is very poor and not ok.
This is corrected.

Fixed nullptr access in destructor of Playback and Capture instances.
This happens when these objects are created and released

Added more control towards capture and dumping the captured data.

Bug: 259219501
Test: atest audiorecord_tests audiotrack_tests audioeffect_tests

Change-Id: I33f404e1ba1c447df802e3a91dd30f53a9abf6b5