ESQueue: Return MIME type with version parameter for AC-4 L4
Parse the AC-4 TOC and construct the AC-4 MIME type with the version
parameter when the first presentation is level 4 or higher. MIME type is
in the following format:

    audio/ac4;version=xx.yy.zz

where xx is the bitstream version, yy is the presentation version and zz
is the presentation level.

Return the base AC-4 MIME type string for legacy presentation levels
less than level 4. Return MIME type string via the kKeyMIMEType key.

Test: The following test passes:
      $ atest CtsMediaExtractorTestCases: \
            android.media.extractor.cts.MediaExtractorTest# \
                testGetAudioPresentations

Change-Id: I0c0bbd6f2b2f3ad5021c2606d919a2216b2f9804