ESQueue: Return AC-4 bitstream version information via kKeyCodecsString
Parse the AC-4 TOC and construct codecs string for the kKeyCodecsString
key that maps to the MediaFormat.KEY_CODECS_STRING key. The version
is based on the first presentation present in the bitstream.

The codecs string is in the following format:

    codecs=ac-4.xx.yy.zz

where xx is the bitstream version, yy is the presentation version and zz
is the presentation level.

Test: The following test passes:
      $ atest CtsMediaExtractorTestCases: \
            android.media.extractor.cts.MediaExtractorTest# \
                testGetAudioPresentations

Change-Id: I0c0bbd6f2b2f3ad5021c2606d919a2216b2f9804