Revert "Spatializer: Log latency requests to/from output sink"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78c2a1b8-b002-4eae-9679-e5e79bdf5137). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If838fbd3161c5062af96e2f89ab0527a6a94a455