OggExtractor#findPrevGranulePosition() only need to read 5000 bytes at most each time back up
When play a dirty ogg file (such as the front part is ogg content and all the reset of it is mp3 content), OggExtractor#findPrevGranulePosition() will enter a horrible loop to repeat backing up and reading.
It takes long time to read the same data over and over again, actually it only need to read new 5000 bytes at most each time back up.

Bug: 263074833
Test: Playback the attachment ogg file of the Bug.
Change-Id: I67de85bfd5af8301ecd619594a935417f69cc007