Audio: Bugfix: patchTrack and patchRecord not released after screen record
1.Sometimes unregisterPolicyMixes() is called after AudioRecord::stop(),
due to binder delay(Mediaprojection and Audiorecord are in different threads).
If remote submix output device is disconnectd in stopInput() before
entering unregisterPolicyMixes(), updateSecondaryOutputs() will not
be called. It causes that patchRecord and patchTrack cannot be destroyed
after screen recorder stops. So we need call checkSecondaryOutputs()
again in unregisterPolicyMixes() to make sure secondary outputs in
correct status.
2.fix 'Pure virtual function called!'
add lock to avoid patchRecord and patchTrack destroyed in interceptBuffer(),
when patchTrack is not destroyed in time.

Test: screen capture app which uses AudioPlaybackCaptureRecord

Signed-off-by: wudongzhe <wudongzhe@xiaomi.com>
Change-Id: Ibfc01fae8e2c9592758992b5cb9b1712a175f17d