audioflinger: Generate patch handle in RecordThread
Patch handles are supposed to be generated by the audio
framework, to ensure their uniqueness across all HAL modules.
For some historical reason, the framework code may ask
the HAL to generate a handle, however this is unsafe.
Fix RecordThread::createAudioPatch_l to generate a handle
in case when it's not provided by the caller.

Bug: 269551518
Test: atest audiorecord_tests
Change-Id: I9e9d96f1981e492f8bd6508cd3c4f6eb7992aafd