media: get JNI env only once per process
We get JNI env every time we create MediaExtractors but if this fails, we
used to print an error log each time. Since JNI environment is static per
process, only get it once.

Also, lower LOG level to debug as this is not in and of itself an error
(this is expected for MediaExtractor).

Bug: 224668846
Change-Id: I961f48aabb01643c3c225fb586ab722d84890462