C2SoftHevcEnc: Update configured level to min level
The plugin allows level values that are higher than configured values
even when not needed. This would result in bitstreams with level
information set to the highest even for basic config like qcif, cif, ...
Worst if the level is too high it can result in bitstreams not being
decodable by the decoders present on the device.

Bug:
Test: atest android.mediav2.cts.EncoderProfileLevelTest

Change-Id: I670eeda9fd9ef6124d1318d493943fc4f803f9be