Revert "Fix AudioSystemTest#StartAndStopAudioSource flakiness"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/289730e5-1b13-4a6e-b230-8aeea6f97f0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If19a30023850b61bde61532f241b4b4147aed3ed