Revert "Fix AudioSystemTest#StartAndStopAudioSource flakiness"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a08d21e5-184d-4d62-88d4-25fed9963fb0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c8e04492ffa1437a6834addd9c0d54ca600b755