Revert "Fix AudioSystemTest#StartAndStopAudioSource flakiness"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25f941ca-4476-4396-8a0e-b4ff36d5bb6e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I570d2bd9c18761d193fa91d67a5791048a9c03dd