Revert "Fix AudioSystemTest#StartAndStopAudioSource flakiness"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8619a7c5-015c-4d2a-80cf-deec691212fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I57aa8a70fa8e8289997b46b5caa5ab6775ca0fb5