Revert "Fix mSession NULL dereference"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc70bf7c-423f-412d-91ba-d1c369167ebd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a37e0a0c12be43f4d2063b614f6eafaf58eb4e3