Revert "Fix build log spam for libmediadrm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b46b8120-5a45-4761-bd20-b8a390bab1c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I14dd7ac0ff1d53115fdd30cdddbe37ad5980a4cf