C2SoftMpeg4Enc: Let encoder library choose apt profile and level
Encoder libraries tends to adjusts the level information configured
to the necessary value if the configured value does not comply with
encoding settings. But aosp mpeg4/h263 enc fails from instantiation
when the configured profile and level information is not in accordance
with the encoding settings.

So, always configure the max profile and level supported and let
the library choose an apt profile and level to place in the bitstream.

Also clip the configured level to max supported level

Bug: 277528543
Test: atest android.media.codec.cts.MediaCodecTest#testVendorParameters \
      android.mediav2.cts.EncoderProfileLevelTest

Change-Id: I7a9359ff4417c36caa7780e30f558cadfb7289e7