Revert "Move EffectsConfig loading in AudioPolicyEffects to libA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0ed458e-d06b-43bd-aeb6-8f5d735bc5d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I40741dc1539fd64f409a4a9658d30082c4d8c26b