Revert "Move EffectsConfig loading in AudioPolicyEffects to libA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d4ba3c56-8979-4c97-b601-caf0b6d14507). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5b5391e5370d3c54d1549bdeeb107120459e386