Revert "Move EffectsConfig loading in AudioPolicyEffects to libA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bdd43148-6326-446a-a293-a99e600ce65e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I14e7ad2ca52c5e2144a6ef60c5ef6ca3ad99cec0