Include some shared libs to static in audioserver
These libraries are only used in this one module so there is no need to
include them as shared libraries.

Including these as shared libraries cost extra storage space and extra
run time memory.

There is a side effect of needing to include the libraries' shared
dependencies directly in the audioserver module.

Test: Boot cuttlefish and:
$ adb shell procrank | grep audioserver
$ adb shell 'showmap -a $(pidof audioserver)'
$ ls -l system/lib64/libaaudioservice.so
$ ls -l system/bin/audioserver
Test: Same test with Pixel 7 device

Bug: 280829178
Change-Id: Id500f63b0415f7cf6fabd9a9d915b0def8127b5a