Fix up mediaswcodec security policies.
Add missing include of crash_dump policies for arm/arm64/riscv64
so that updates to crash_dump/debuggerd happen automatically.
Remove the old section from the policy that duplicated the crash_dump
information.

Add an explicit section for AddressSanitizer syscalls to the files
missing them to match all of the other swcodec policy files.

Change the mprotect and mmap rules to just allow those syscalls
with any arguments. The rule only really disallowed an argument
that contained both PROT_EXEC and PROT_WRITE. If it was only
PROT_EXEC, it would be pass the rule. The rules inherited from
crash_dump are more restrictive and disallow PROT_EXEC completely.
Without this change, the minijail parser fails because it doesn't
appear to like multiple arg related rules for the same syscall.

Test: Ran debuggerd <PID_OF_media.swcodec> on 64 bit arm device.
Test: Ran debuggerd <PID_OF_media.swcodec> on 32 bit arm device.
Change-Id: I33704daa92751537a0d1f277055b2118d2d249e5