WIP Convert MediaBenchmark to use MediaPreparer
This is WIP because there's something about the adb install that is
wonky and it appears to be a flag new to U:
IllegalArgumentException: Unknown option --bypass-low-target-sdk-block

I'm doing this with mixed adb, an ssh tunnel to another box, and aosp
atest. Plenty of opportunities for me to have something out of whack.

I'll come back to this soon with a cleaner configuration.

Bug: 274674920
Test: atest MediaBenchmarkTest
Change-Id: Icd2dc9d03b980d91b90032c99a783f00933a8a3e