Revert "SyncEvent: modernize C++"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bafb0195-16b4-4f49-963e-9f87c0977667). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie943f7d05b6a4e0f8e340ba4d014f8de40680702