Revert "AudioFlinger: move fast path components to library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd4a70e9-1a30-4f08-a130-f454edddcd75). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55866d4767a9c7f98d9446b1ec3175ebc6ca4d19