Revert "implementing concurrent codec metrics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e326738a-5ebc-4d3f-bd0b-4b5e9771f28f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9fb7288184595554121bd134d8753a1df4c12f90