Revert "Reduce logging verbosity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/403998c2-446c-43a1-99c2-388a3f90f0c6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75213a5389da9241fb0ee00179790ca9e103edcf