Revert "resourcemanagerservice: make a copy of client config"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/beeb60aa-4f17-4248-b6bd-cfb536895a55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1009fd28dcb1a66bfdbb44677d9a73ca1b2d7d5b