Revert "media: update on concurrent codec usage"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/beeb60aa-4f17-4248-b6bd-cfb536895a55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3f7778e3a533c3d6de0425a43d13cfbd766379d