Revert "codec2 hal: type conversion refactoring, step 2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7aa9011-6e1a-4bc9-98b4-0ab7c29a68d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib40e71b98e89b62300277de837a6c81b26cd8303