Revert "AIDL: Consider both inputs and output of r_submix to be ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c366f501-9ef8-432e-abdb-c369f54919a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a1ea1ce0a24ad03564fcd596306cb0cff87ea8b