Revert "codec2 hal: example AIDL service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd94fae3-7415-40dd-b0d3-612906079279). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c38912295211f69c872f3f97ba9a0f4e9386ad1