Revert "codec2 hal: fuzzer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd94fae3-7415-40dd-b0d3-612906079279). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ed8832de33052a08f9b807cdecfbc797b6f2e7b