Revert "codec2 hal: example AIDL service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f11a542-fc29-4d68-8a5e-a09d1d18fac8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0540a4e578a9b9af871cb685ab4f5b959c9bed9