resourcemanager: resource model and reclaim policy
Refactor the current design of the ResourceManagerService to
introduce Resource Model and Reclaim Policies.
This defines Resource Model as an interface and
implements the default model that current implementation uses.

Since this is a refactoring change without the addition of
new API to the ResourceManagerService, it is unflagged.
This change allows extending resource models and reclaim policies.

Bug: 289097671
Test: atest android.media.misc.cts.ResourceManagerTest
      atest android.media.misc.cts.ResourceManagerMultiTest
      /data/nativetest64/ResourceManagerService_test/ResourceManagerService_test
      /data/nativetest64/ResourceObserverService_test/ResourceObserverService_test
Change-Id: I0b962ea79e6e5e63baba3050718ebd415e68f5a1