Revert "Replace error_occurred with skipped"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/abc48e4c-0418-407c-9764-0ea68d42f465). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0453100d1fbd4e505553774ccdcc6deb9459c7f