resourcemanager: define reclaim policy
Refactor the current design of the ResourceManagerService to
introduce Reclaim Policies.
This defines and implements the priority (oom score) of the process
as the default reclaim policy.

Since this is a refactoring change without the addition of
new API to the ResourceManagerService, it is unflagged.
This change allows extending resource models and reclaim policies.

Bug: 294886363
Test: atest android.media.misc.cts.ResourceManagerTest
      atest android.media.misc.cts.ResourceManagerMultiTest
      /data/nativetest64/ResourceManagerService_test/ResourceManagerService_test
      /data/nativetest64/ResourceObserverService_test/ResourceObserverService_test
Change-Id: I0de0d86c08b1b148e2f813b884aa693b1643900c