Revert "C2SoftDav1dDec: Move libyuv calls to SimpleC2Component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c8661cf-51ea-413e-9eb4-695409119528). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic9ec8185d2a48c6c8b2660a33ebcd439d6d5266a