Revert "Support dumping from given position and migrate code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c8661cf-51ea-413e-9eb4-695409119528). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib38f023e0a0fe59a5fd14f49f1ac830b27461ea9