Revert "Support the dav1d decoder in a new codec2 component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c8661cf-51ea-413e-9eb4-695409119528). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12bbb04ffe1f26386067316fcdc17a1c42807067