Revert "C2SoftDav1dDec: Replace memcpy with dav1d_data_wrap"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c8661cf-51ea-413e-9eb4-695409119528). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a2d6f8be200e892c3185eae173d1ac3c9625e50