Revert "C2SoftDav1dDec: Replace memcpy with dav1d_data_wrap"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fee51331-0f3e-45bd-98cb-99f5f10c2a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12c2821889d0affd7f9f0c5e5e2615accb1de1c7