Revert "C2SoftDav1dDec: Move libyuv calls to SimpleC2Component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fee51331-0f3e-45bd-98cb-99f5f10c2a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8100a73dd39ab254dc3dc291f879b74f15f7459