Revert "Support dumping from given position and migrate code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fee51331-0f3e-45bd-98cb-99f5f10c2a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9971ae80a1bbe5eff463d45cede60567ff81155