Revert "Support the dav1d decoder in a new codec2 component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fee51331-0f3e-45bd-98cb-99f5f10c2a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5557c4dcd5f16c04b086e839543222c5c0770b1