Revert "Support dumping from given position and migrate code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/234f577f-d994-4d9b-b017-1303c232dc1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1ab3e71e6e9680d3f675e1fabf2edf32a9ebe34