Revert "Support the dav1d decoder in a new codec2 component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/234f577f-d994-4d9b-b017-1303c232dc1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I441a330ff03f6dfe93d2f5366091755a662bc0d7