Revert "C2SoftDav1dDec: Replace memcpy with dav1d_data_wrap"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/234f577f-d994-4d9b-b017-1303c232dc1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4132e321eb1411a2a11b3b7abdbceccaba15e1f