Revert "C2SoftDav1dDec: Move libyuv calls to SimpleC2Component"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/234f577f-d994-4d9b-b017-1303c232dc1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied3435bb1ee16a6781f7b0e87e173e3cc9600b25