CCodecConfig: Fix issue where ts-schema isn't returned if no b-layers configured
ts-schema may be configured with only temporal layer count,
without b-layers. For example, android.generic.2 is a legitimate
ts-schema without b-layers. However, CCodecConfig::getFormatForDomain
didn't set ts-schema when m.bLayerCount is 0.
This caused an issue that even if ts-schema is configured,
it's not returned in MediaCodec.getOutputFormat.

Fix by checking m.ㅣayerCount if m.bLayerCount is 0.

Test: manual
Bug: https://issuetracker.google.com/issues/306222291
Change-Id: I72329c17265bd817518e5b2028296688bf3e4cd7