Revert "Enable dav1d decoder as another av1 decoder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9310a213-07d6-47fe-beda-305b755adcd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I400ca55574fd13fca39e2abba595ed29a748560a