audiopolicy/managerdefault: Allow non-attached default output devices
Some embedded devices, like set-top boxes, may not have attached
output audio devices at all. And it's common for them to have an HDMI
port (and maybe Bluetooth) as the only option to output audio. And
it's natural for them to use a non-attached output device as the
defaultOutputDevice.

Unfortunately, the default APM doesn't allow this. To initialize APM,
the defaultOutputDevice should be available. For non-attached devices,
it means that the device state should be 'connected', but it's
impossible to change the device connection state, because APM is not
initialized yet.

Until Android 10, the result of AndroidPolicyManager::initialize() was
simply ignored. So it was possible to use HDMI (and other non-attached
output devices) as the defaultOutputDevice.

The breaking change was introduced in commit caafba4078 ("audio
policy: fix virtual called from constructor").

Test: manual - set non-attached device as defaultOutputDevice and run
stop audioserver && start audioserver

Change-Id: Idd2ad960d80ae3587c162fd75cfb333de47183ec
Signed-off-by: Sergey Bobrenok <SIBobrenok@sberdevices.ru>