Revert "Add media.c2 AIDL service to media.swcodec"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d529ee64-5dc7-49a2-9ea1-6ffbb096b859). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a0378da67dd0306a86de247893e10deeb31e760