Revert "audiopolicy_fuzzer.cpp : Bug fix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8ef7b2ca-b22a-47e4-8ddc-c449f33c5b0b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id037053be121fc1519ac95fc963c34722c25f399