Revert "Ignore SIGPIPE in codec2 AIDL fuzzer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0008e53f-bd49-41ae-a888-9bc09904a94d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1dea50700e5d30b38b48c60d9405f7e35f995925