libaudiohal@aidl: Properly handle "legacy" r_submix case
The "legacy" r_submix case is the "default" remote submix
device with the address "0". For backwards compatibility,
the input for this device must be attached. The output
is still opened dynamically. However, the framework expects
to see a declared device with address "0" and some profiles.

Added code to libaudiohal@aidl to perform this adaptation.
Also, found that `DeviceDescriptor` was missing setting
`mDeclaredAddress` when loading from a parcelable.

Bug: 302036943
Test: atest audioeffect_analysis
Test: atest CtsMediaAudioTestCases --test-filter=".*AudioPlaybackCaptureTest.*"
Change-Id: I28294cbb2895e648fe962feb3e9a0645fc823153