Fix playback thread processing format for spdif out
The processing format was set to the HAL format, rather than the raw
encoded format, for cases when the SpdifStreamOut path is taken.

Also assign mSampleRate and mChannelMask (by definition, HAL values)
from the mOutput stream directly, rather than the mOutput (wrapper) as
for the SpdifStreamOut case, the mOutput wrapper audio properties are
the application facing ones and would be incorrect to assign to the HAL
mSampleRate / mChannelMask (this is currently benign, because
SpdifStreamOut does not override getAudioProperties), but it's
confusing, so improve clarity.

Change-Id: Iaca53f823b862ead135dd4820e232098aed605c7