Fix playback thread processing format for spdif out
The processing format was set to the HAL format, rather than the raw
encoded format, for cases when the SpdifStreamOut path is taken.

Also assign mSampleRate and mChannelMask (by definition, HAL values)
from the mOutput stream directly, rather than the mOutput (wrapper) as
for the SpdifStreamOut case, the mOutput wrapper audio properties are
the application facing ones and would be incorrect to assign to the HAL
mSampleRate / mChannelMask (this is currently benign, because
SpdifStreamOut does not override getAudioProperties), but it's
confusing, so improve clarity.

Test: dumpsys shows the properties for an E-AC3-JOC output stream
      wrapped by SpdifStreamOut
  Sample rate: 192000 Hz
  HAL format: 0x1 (AUDIO_FORMAT_PCM_16_BIT)
  Channel count: 2
  Channel mask: 0x00000003 (front-left, front-right)
  Processing format: 0xa000001 (AUDIO_FORMAT_E_AC3_JOC)
  Processing frame size: 1 bytes

Change-Id: Iaca53f823b862ead135dd4820e232098aed605c7