Add a boolean for peek enabled
Peek state is always changed to kEnabledNoBuffer after start()
and flush(), if it was not in legacy mode. This requires app
to always disable peek after start or flush if it wants to.

The fix is Adding a boolean to indicate if peek is enabled and
set the state accordingly in start() and flush().

Test: atest DecodeOnlyTest
Bug: 287456023
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:2dd1835fd2870a609395efad624454af350b73b4)
Merged-In: Ib5a93e9b919515fe9752043e72553022c71b248e
Change-Id: Ib5a93e9b919515fe9752043e72553022c71b248e