libaudiohal@aidl: Align ownership of patches with the framework
DeviceHalAidl creates patches when opening streams.
Later these patches can be "re-created" by the framework,
that means, the framework creates a patch for the same
pair of device and mix port configurations via
'createAudioPatch' call. In this case, ownership of
the patch must be transferred to the framework, it will
release it via 'releaseAudioPatch' later.

Properly separate ownership of port configs between streams
and patches to avoid attempting to release port configs
belonging to patches.

Bug: 302132812
Bug: 303742495
Test: run `atest CtsMediaAudioTestCases \
           --test-filter=".*AudioPlaybackCaptureTest#testPlaybackCaptureDoS`
      and check for errors from the AIDL HAL
Change-Id: I695af339e7293a12d07cf0d35d10a4c29008a3d2