Revert "Restrict unsigned integer overflow in MPEG4Writer writeE..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3a0b3b0-73f4-46f3-831f-09e086daefdb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1493a29e41c9a4b2abb44218c41f0134637ac55a