Revert "Restrict unsigned integer overflow in MPEG4Writer writeE..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a7bcad9-3c01-4961-941a-227b9cc17cc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad49934358a4f54f8e67afb13eb85e8ba8934ebc