Revert "Enable AIDL HAL support in the framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/76041be9-0302-4b80-8f83-e5c86bc84a8a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I648b2c2b5ceb7c9cd49ea5f808399d7b90d04be7