Revert "Enable AIDL HAL support in the framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5abef26-2dca-4ec9-9c8a-8829ff36f47d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I93ef15c1ee07587df00eb132060cafa17a5e800e