libaudiohal@aidl: Improve handling of unsupported mix port configs
Sometimes the framework provides a mix port config which does not
correspond to any of profiles (likely, it's the client app config).
When it happens on stream opening, try to use a device port config.
When it happens on patch setting, try finding an existing port
config using flags and I/O handle.

Bug: 264712385
Bug: 315528763
Test: atest CtsVirtualDevicesTestCases --test-filter=".*VirtualAudioTest.*"
Change-Id: I172904411b27c20fc2d2247d5b9b09b4451c5c34