libaudiohal@aidl: Fix handling of stream config suggestion
The contract of 'DeviceHalInterface::open{Input|Output}Stream'
states that in the case when the HAL suggests a config which is
different from the provided one, the method must return 'BAD_VALUE'
and fill out the suggested config. The framework then may make
another attempt using the suggested config. This behavior was
missing in libaudiohal@aidl prior to this change.

Bug: 264712385
Test: atest CtsMediaAudioTestCases
Change-Id: I321b9701473384902bbe533607abb137fb8a07cd