Revert "libaudiohal@aidl: Work around use of unsupported mix por..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b2a2b8d-d50e-492a-a77a-7f1235777342). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7829134eb08e880197645cc7bb88d9a79c35cc1