[DTV04155345][google_path]Add a patch for KEY_PRIORITY
[Description]
This is a upstream patch for KEY_PRIORITY modification.

[Root Cause]
CCodecConfig alwlays invert the value of KEY_PRIORITY from positive to negative.
This way, Codec2 need to invert it again.
Otherwise, the ranking mechanism is wrong.

[Solution]
Remove the inversion method in the CCodecConfig.cpp.

[Test Report]
Build successfuly, and run the image normally on the mokona platform.

CR-Id: DTV04155345
Change-Id: I1637b510d5827e8f382d45d00488cd53775e87d6