Revert "Align A2DP and speaker volume in tandem case"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a8b48aa5-17e2-412d-8b93-7f2518fa9a79). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I06404abfe9c19826e146d3e07e1ec94f1d599ab6