[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
<add g3 doc link>

Bug: 304529413
Test: N/A
Change-Id: Iff4eca6c665999ee1fcb2caa1d0a6608c775ead3
Merged-In: Iff4eca6c665999ee1fcb2caa1d0a6608c775ead3