Update max limit for indices used when enumerating supported formats and profiles.
A cap of 32 was introduced on the number of profile-level pairs to
check for each codec to avoid infinite loops. This limit is preventing
the AVC codec from reporting its upper limit support for High Profile.

Based on
https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/media/java/android/media/MediaCodecInfo.java
MPEG4 supports up to 16 profiles and 10 levels making it possible for
capable codecs to go up to 160 indexes. As a solution, allow the worst
case number of combinations of profile and levels to be checked.

Test: Build is successful, atest MediaCodecCapabilitiesTest, atest
CtsVideoTestCases, atest mediacodecTest, atest CtsMediaMiscTestCases,
atest CtsMediaPlayerTestCases

Change-Id: Iaa222878458562b44f5bfc7f7d06f73fc938de75
Signed-off-by: Saurabh <sbbanore@amazon.com>